added removeInputField. fixed bugs
